requirements.txt
faiss-cpu
google-genai
sentence-transformers
streamlit
numpy
python-dotenv
pymupdf
pytesseract
Pillow
